perf: sweep 10 - sqrtf/expf/cosf/sinf/fabsf on float data

+++ plugins/milk-extra-src/image_basic/imgetcircsym.c
@@ -180,7 +180,7 @@ imageID IMAGE_BASIC_get_circsym_component(
         for(uint32_t ii = 0;
             ii < naxes[0]; ii++)
         {
-            float distance = sqrt(
+            float distance = sqrtf(
                 (1.0 * ii - xcenter)
                 * (1.0 * ii - xcenter)
                 + (1.0 * jj - ycenter)
@@ -237,7 +237,7 @@ imageID IMAGE_BASIC_get_circsym_component(
         for(uint32_t ii = 0;
             ii < naxes[0]; ii++)
         {
-            float distance = sqrt(
+            float distance = sqrtf(
                 (1.0 * ii - xcenter)
                 * (1.0 * ii - xcenter)
                 + (1.0 * jj - ycenter)
